refactor: docs example
